+++ CHANGELOG.md
@@ -8,6 +8,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 ### Fixed
 - Put `vcs`, `vsim`, and `riviera` defines in quotes.
 
+### Added
+- Add flag for `rtl` target to files without target in script and sources.
+
 ## 0.28.1 - 2024-02-22
 ### Added
 - Add `flist-plus` script format for file list with plusargs.